move precompile test to generic
